python/example_code/bedrock-agent-runtime/code-interpreter/requirements.txt
boto3>=1.28.0
matplotlib>=3.7.0
pillow>=9.4.0
